# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_sayHi","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/local/signup":{"post":{"operationId":"AuthController_signupLocal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":"Signup succeeded / Created successfully"},"400":{"description":"Bad Request, doesn't match the Dto blue-print"},"403":{"description":"Signup failed"}},"tags":["Auth"]}},"/auth/local/signin":{"post":{"description":"You must provide either **email** or **phone**. If `email` is missing, `phone` is required, and vice versa.","operationId":"AuthController_signinLocal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"Signup succeeded"},"400":{"description":"Bad Request, doesn't match the Dto blue-print"},"401":{"description":"Forbidden!, wrong email or password"},"403":{"description":"Signup failed"}},"summary":"Sign in using email or phone","tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[{"name":"Authorization","in":"header","description":"Should be like this: 'Bearer access_token'","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":"Logged out successfully"},"401":{"description":"Unauthorized (expired access_token)"}},"tags":["Auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshTokens","parameters":[{"name":"Authorization","in":"header","description":"Should be like this: 'Bearer refresh_token'","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":"AccessToken was refreshed successfully"},"403":{"description":"User Should be Logged Out. not-exit || expired-RT || banned"}},"tags":["Auth"]}},"/auth/verify":{"post":{"operationId":"AuthController_verifyAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifingDto"}}}},"responses":{"403":{"description":"verification code wrong or expired"}},"tags":["Auth"]}},"/auth/resend-verification-code":{"post":{"operationId":"AuthController_resendVerificationCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserIdDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/request-reset-password":{"post":{"operationId":"AuthController_requestResetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserIdDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}}},"info":{"title":"Talabaty Backend Documentation","description":"Endpoints' description","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignupDto":{"type":"object","properties":{"name":{"type":"string","description":"username, should be 3-30 characters"},"email":{"type":"string","description":"any valid email address"},"phone":{"type":"string","description":"string","example":"+963 911111111, 0911111111"},"password":{"type":"string","description":"6-30 charachters long password","example":"123456"}},"required":["name","email","phone","password"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","description":"Email of the user. Required if phone is not provided.","example":"user@example.com"},"phone":{"type":"string","description":"Phone of the user. Required if email is not provided.","examples":["+963 911111111","0911111111","911111111"]},"password":{"type":"string"}},"required":["password"]},"VerifingDto":{"type":"object","properties":{"userId":{"type":"string","description":"user's id","example":"eo8179-exlholr-slklkjl..."},"code":{"type":"number","description":"4 digits integer","example":"0000"}},"required":["userId","code"]},"UserIdDto":{"type":"object","properties":{"userId":{"type":"string"}},"required":["userId"]},"ResetPasswordDto":{"type":"object","properties":{"userId":{"type":"string","description":"user's id, to put extra protection to the endpoint!","example":"lkjlsdf-asdlkl-lkjli...."},"code":{"type":"number","description":"the verification-code sent to email","example":12345},"email":{"type":"string","description":"valid email, we added it also to put extra protection","example":"aaa@example.com"},"newPassword":{"type":"string"}},"required":["userId","code","email","newPassword"]}}}}